fix(core): stop int overflow in scoring and validate quality-ladder cutoff
- GroupScore and TermScore in ReleaseScorer now accumulate into long
  (Score already did). A large enough user-supplied preference score
  multiplied by GroupScoreScale/TermScoreScale wrapped a maximally
  preferred group into a maximally discouraged one in int arithmetic,
  defeating the soft-score clamp that was added to make quality-step
  dominance structural.
- QualityLadder now validates CutoffName against its rungs at
  construction and throws ArgumentException naming the bad value and
  the available rung names. CutoffRank's int.MaxValue fallback stays
  as defence in depth, but a mistyped cutoff no longer silently turns
  MeetsCutoff false for every quality.
- Tighten Score_HandlesAQualityNotOnTheLadderWithoutThrowing (renamed
  Score_TreatsAnOffLadderQualityAsTheBottomRung) to assert the actual
  score lands in the rank-0 band instead of only asserting it does not
  throw.

+++ src/NoMercy.Plugin.TorrentDownloader.Core/Profiles/ReleaseScorer.cs
@@ -47,31 +47,31 @@ public class ReleaseScorer
             : 0;
     }
 
-    private static int GroupScore(ParsedRelease parsed, ReleaseProfile profile)
+    private static long GroupScore(ParsedRelease parsed, ReleaseProfile profile)
     {
         if (parsed.ReleaseGroup is not string group)
             return 0;
 
-        int total = 0;
+        long total = 0;
         foreach (GroupPreference preference in profile.PreferredGroups)
         {
             if (string.Equals(preference.Group, group, StringComparison.OrdinalIgnoreCase))
-                total += preference.Score * GroupScoreScale;
+                total += (long)preference.Score * GroupScoreScale;
         }
 
         return total;
     }
 
-    private static int TermScore(string title, ReleaseProfile profile)
+    private static long TermScore(string title, ReleaseProfile profile)
     {
-        int total = 0;
+        long total = 0;
         foreach (TermRule term in profile.Terms)
         {
             if (term.Kind != TermKind.Preferred)
                 continue;
 
             if (TermMatcher.IsMatch(title, term.Pattern))
-                total += term.Score * TermScoreScale;
+                total += (long)term.Score * TermScoreScale;
         }
 
         return total;
